Cleanup mamorx service
