+++ Mamorx/services/mamorx_service/app/helper/utils.py
@@ -1,10 +1,10 @@
 import uuid
 import base64
-from datetime import datetime
 
-from app.schemas import ReviewJob, ReviewType
+from MAMORX.schemas import ReviewJob, ReviewType
 
 def create_review_job(
+        filename: str,
         pdf_data: bytes,
         session_id: str,
         review_type: ReviewType
@@ -14,6 +14,7 @@ def create_review_job(
     job_to_submit = ReviewJob(
         id=id,
         session_id=session_id,
+        filename=filename,
         review_type=review_type,
         pdf_content=base64.b64encode(pdf_data)
     )
